Extracting bread matrix for singleR class

Extracting bread matrix for singleR class

Usage

# S3 method for singleR
bread(object, ...)

Arguments

object

an object representing a fitted model.

...

additional optional arguments passed to the following functions:

stats::vcov -- for extracting "the usual" variance-covariance matrix, vcov.singleR has one additional argument type with values "Fisher" "observedInform", defaults to the one specified in controlPopVar specified in call for object.

Value

A bread matrix, i.e. a hessian based estimation of variance-covariance matrix scaled by degrees of freedom.

See also

Author

Piotr Chlebicki, Maciej Beręsewicz

An S3class for sandwich::bread to handle singleR objects. This function was developed based on sandwich:::bread.glm

Examples

Model <- estimatePopsize(
formula = capture ~ ., 
data = netherlandsimmigrant, 
model = ztpoisson, 
method = "IRLS")
sandwich::bread(Model)
#>                      (Intercept) gendermale   age>40yrs reasonOther reason
#> (Intercept)            122.10868 -34.311532 -10.9034770        -41.6072585
#> gendermale             -34.31153  49.979685  -2.5603625         -1.0221200
#> age>40yrs              -10.90348  -2.560362 313.3232550          0.1153169
#> reasonOther reason     -41.60726  -1.022120   0.1153169         49.0525800
#> nationAsia             -56.25448  -6.805030   6.9235321         -0.8766691
#> nationNorth Africa     -55.61099  -9.030616   8.2485110          0.1373466
#> nationRest of Africa   -60.88260  -4.806497   4.5385472          2.9626704
#> nationSurinam          -60.40757  -2.801890  -3.5269990          1.1476629
#> nationTurkey           -51.41920  -9.448466  -2.1998382         -3.8913678
#>                       nationAsia nationNorth Africa nationRest of Africa
#> (Intercept)          -56.2544755        -55.6109935           -60.882602
#> gendermale            -6.8050296         -9.0306164            -4.806497
#> age>40yrs              6.9235321          8.2485110             4.538547
#> reasonOther reason    -0.8766691          0.1373466             2.962670
#> nationAsia           171.0514100         62.9320747            62.223177
#> nationNorth Africa    62.9320747         70.7547976            62.510353
#> nationRest of Africa  62.2231766         62.5103529           170.291804
#> nationSurinam         61.8198398         61.9519618            61.820909
#> nationTurkey          62.8528821         63.1825931            62.171390
#>                      nationSurinam nationTurkey
#> (Intercept)             -60.407570   -51.419195
#> gendermale               -2.801890    -9.448466
#> age>40yrs                -3.526999    -2.199838
#> reasonOther reason        1.147663    -3.891368
#> nationAsia               61.819840    62.852882
#> nationNorth Africa       61.951962    63.182593
#> nationRest of Africa     61.820909    62.171390
#> nationSurinam          1931.372466    62.008295
#> nationTurkey             62.008295   683.378471
vcov(Model)
#>                       (Intercept)    gendermale     age>40yrs
#> (Intercept)           0.064951423 -0.0182508147 -5.799722e-03
#> gendermale           -0.018250815  0.0265849388 -1.361895e-03
#> age>40yrs            -0.005799722 -0.0013618949  1.666613e-01
#> reasonOther reason   -0.022131520 -0.0005436809  6.133879e-05
#> nationAsia           -0.029922593 -0.0036196966  3.682730e-03
#> nationNorth Africa   -0.029580316 -0.0048035194  4.387506e-03
#> nationRest of Africa -0.032384363 -0.0025566472  2.414121e-03
#> nationSurinam        -0.032131686 -0.0014903669 -1.876063e-03
#> nationTurkey         -0.027350636 -0.0050257797 -1.170127e-03
#>                      reasonOther reason    nationAsia nationNorth Africa
#> (Intercept)               -2.213152e-02 -0.0299225934      -2.958032e-02
#> gendermale                -5.436809e-04 -0.0036196966      -4.803519e-03
#> age>40yrs                  6.133879e-05  0.0036827298       4.387506e-03
#> reasonOther reason         2.609180e-02 -0.0004663133       7.305668e-05
#> nationAsia                -4.663133e-04  0.0909847926       3.347451e-02
#> nationNorth Africa         7.305668e-05  0.0334745078       3.763553e-02
#> nationRest of Africa       1.575889e-03  0.0330974344       3.325019e-02
#> nationSurinam              6.104590e-04  0.0328828935       3.295317e-02
#> nationTurkey              -2.069876e-03  0.0334323841       3.360776e-02
#>                      nationRest of Africa nationSurinam nationTurkey
#> (Intercept)                  -0.032384363  -0.032131686 -0.027350636
#> gendermale                   -0.002556647  -0.001490367 -0.005025780
#> age>40yrs                     0.002414121  -0.001876063 -0.001170127
#> reasonOther reason            0.001575889   0.000610459 -0.002069876
#> nationAsia                    0.033097434   0.032882893  0.033432384
#> nationNorth Africa            0.033250188   0.032953171  0.033607762
#> nationRest of Africa          0.090580747   0.032883462  0.033069888
#> nationSurinam                 0.032883462   1.027325780  0.032983136
#> nationTurkey                  0.033069888   0.032983136  0.363499187
# This function just scales.
all(vcov(Model) * nrow(netherlandsimmigrant) == sandwich::bread(Model))
#> [1] TRUE
# We can choose Fisher information matrix instead of default observed information matrix.
vcov(Model, "Fisher")
#>                       (Intercept)    gendermale     age>40yrs
#> (Intercept)           0.064951423 -0.0182508146 -5.799722e-03
#> gendermale           -0.018250815  0.0265849387 -1.361895e-03
#> age>40yrs            -0.005799722 -0.0013618949  1.666613e-01
#> reasonOther reason   -0.022131520 -0.0005436809  6.133875e-05
#> nationAsia           -0.029922593 -0.0036196966  3.682730e-03
#> nationNorth Africa   -0.029580316 -0.0048035193  4.387506e-03
#> nationRest of Africa -0.032384363 -0.0025566472  2.414121e-03
#> nationSurinam        -0.032131686 -0.0014903669 -1.876063e-03
#> nationTurkey         -0.027350636 -0.0050257797 -1.170127e-03
#>                      reasonOther reason    nationAsia nationNorth Africa
#> (Intercept)               -2.213152e-02 -0.0299225934      -2.958032e-02
#> gendermale                -5.436809e-04 -0.0036196966      -4.803519e-03
#> age>40yrs                  6.133875e-05  0.0036827298       4.387506e-03
#> reasonOther reason         2.609180e-02 -0.0004663133       7.305668e-05
#> nationAsia                -4.663133e-04  0.0909847926       3.347451e-02
#> nationNorth Africa         7.305668e-05  0.0334745078       3.763553e-02
#> nationRest of Africa       1.575889e-03  0.0330974344       3.325019e-02
#> nationSurinam              6.104590e-04  0.0328828935       3.295317e-02
#> nationTurkey              -2.069876e-03  0.0334323841       3.360776e-02
#>                      nationRest of Africa nationSurinam nationTurkey
#> (Intercept)                  -0.032384363  -0.032131686 -0.027350636
#> gendermale                   -0.002556647  -0.001490367 -0.005025780
#> age>40yrs                     0.002414121  -0.001876063 -0.001170127
#> reasonOther reason            0.001575889   0.000610459 -0.002069876
#> nationAsia                    0.033097434   0.032882893  0.033432384
#> nationNorth Africa            0.033250188   0.032953171  0.033607762
#> nationRest of Africa          0.090580747   0.032883462  0.033069888
#> nationSurinam                 0.032883462   1.027324675  0.032983136
#> nationTurkey                  0.033069888   0.032983136  0.363499187
sandwich::bread(Model, type = "Fisher")
#>                      (Intercept) gendermale   age>40yrs reasonOther reason
#> (Intercept)            122.10868 -34.311531 -10.9034769        -41.6072583
#> gendermale             -34.31153  49.979685  -2.5603625         -1.0221200
#> age>40yrs              -10.90348  -2.560362 313.3232522          0.1153168
#> reasonOther reason     -41.60726  -1.022120   0.1153168         49.0525798
#> nationAsia             -56.25448  -6.805030   6.9235321         -0.8766691
#> nationNorth Africa     -55.61099  -9.030616   8.2485110          0.1373466
#> nationRest of Africa   -60.88260  -4.806497   4.5385472          2.9626704
#> nationSurinam          -60.40757  -2.801890  -3.5269989          1.1476629
#> nationTurkey           -51.41920  -9.448466  -2.1998381         -3.8913678
#>                       nationAsia nationNorth Africa nationRest of Africa
#> (Intercept)          -56.2544756        -55.6109936           -60.882602
#> gendermale            -6.8050296         -9.0306164            -4.806497
#> age>40yrs              6.9235321          8.2485110             4.538547
#> reasonOther reason    -0.8766691          0.1373466             2.962670
#> nationAsia           171.0514100         62.9320747            62.223177
#> nationNorth Africa    62.9320747         70.7547976            62.510353
#> nationRest of Africa  62.2231766         62.5103529           170.291804
#> nationSurinam         61.8198398         61.9519618            61.820909
#> nationTurkey          62.8528821         63.1825931            62.171390
#>                      nationSurinam nationTurkey
#> (Intercept)             -60.407570   -51.419195
#> gendermale               -2.801890    -9.448466
#> age>40yrs                -3.526999    -2.199838
#> reasonOther reason        1.147663    -3.891368
#> nationAsia               61.819840    62.852882
#> nationNorth Africa       61.951962    63.182593
#> nationRest of Africa     61.820909    62.171390
#> nationSurinam          1931.370389    62.008295
#> nationTurkey             62.008295   683.378471
all(vcov(Model, "Fisher") * nrow(netherlandsimmigrant) == sandwich::bread(Model, type = "Fisher"))
#> [1] TRUE