Control parameters specific to some models

controlModel constructs a list with all necessary control parameters in estimatePopsize that are either specific to selected model or don't fit anywhere else.

Specifying additional formulas should be done by using only right hand side of the formula also for now all variables from additional formulas should also be included in the "main" formula.

Usage

controlModel(
  weightsAsCounts = FALSE,
  omegaFormula = ~1,
  alphaFormula = ~1,
  piFormula = ~1
)

Arguments

weightsAsCounts

for now does nothing. The plan is to have this indicate whether prior.weights are to be treated as counts for sub populations and adjust all necessary methods and functionalities, like adjustments in bootstrap or decreasing weights in dfbeta instead or deleting rows from data, to accommodate this form of data.

omegaFormula

formula for inflation parameter in one inflated zero truncated and zero truncated one inflated models.

alphaFormula

formula for dispersion parameter in negative binomial based models.

piFormula

formula for probability parameter in pseudo hurdle zero truncated and zero truncated pseudo hurdle models.

Value

A list with selected parameters, it is also possible to call list directly.

See also

Author

Piotr Chlebicki, Maciej Beręsewicz