Control parameters for population size estimation

Creating control parameters for population size estimation and respective standard error and variance estimation.

Usage

controlPopVar(
  alpha = 0.05,
  trcount = 0,
  bootType = c("parametric", "semiparametric", "nonparametric"),
  B = 500,
  confType = c("percentilic", "normal", "basic"),
  keepbootStat = TRUE,
  traceBootstrapSize = FALSE,
  bootstrapVisualTrace = FALSE,
  fittingMethod = NULL,
  bootstrapFitcontrol = NULL,
  sd = c("sqrtVar", "normalMVUE"),
  covType = c("observedInform", "Fisher")
)

Arguments

alpha

Significance level, 0.05 used by default.

trcount

Truncated count - a number to be added to point estimator and both sides of confidence intervals.

bootType

Bootstrap type. Default is "parametric", other possible values are: "semiparametric" and "nonparametric".

B

Number of bootstrap samples to be performed (default 500).

confType

Type of confidence interval for bootstrap confidence interval, "percentile" by default. Other possibility: "studentized" and "basic".

keepbootStat

Boolean value indicating whether to keep a vector of statistics produced by bootstrap.

traceBootstrapSize

Boolean value indicating whether to print size of bootstrapped sample after truncation for semi- and fully parametric bootstraps.

bootstrapVisualTrace

Boolean value indicating whether to plot bootstrap statistics in real time.

fittingMethod

Method used for fitting models from bootstrap samples.

bootstrapFitcontrol

Control parameters for each regression works exactly like controlMethod but for fitting models from bootstrap samples.

sd

Indicates how to compute standard deviation of population size estimator either as: \[\hat{\sigma}=\sqrt{\hat{\text{var}}(\hat{N})}\] for sqrt or for normalMVUE as the unbiased minimal variance estimator for normal distribution: \[\hat{\sigma}=\sqrt{\hat{\text{var}}(\hat{N})}\frac{\Gamma\left((N_{obs}-1)/2\right)}{\Gamma\left(N_{obs}/2\right)}\sqrt{\frac{N_{obs}}{2}}\] where the ration involving gamma functions is computed by loggamma function.

covType

type of covariance matrix for regression parameters by default observed information matrix, more options will be here in the future.

Value

A list with selected parameters, it is also possible to call list directly.

See also

Author

Piotr Chlebicki, Maciej Beręsewicz