Regression fitting in single source capture-recapture models

estimatePopsize.fit does for estimatePopsize what glm.fit does for glm. It is internally called in estimatePopsize. Since estimatePopsize does much more than just regression fitting estimatePopsize.fit is much faster.

Usage

estimatePopsize.fit(y, X, family, control, method, priorWeights, start, ...)

Arguments

y

vector of dependent variables.

X

model matrix, the vglm one.

family

same as model in estimatePopsize.

control

control parameters created in controlModel.

method

method of estimation same as in estimatePopsize.

priorWeights

vector of prior weights its the same argument as weights in estimatePopsize.

start

initial value of regression parameters.

...

arguments to pass to other methods.

Value

List with regression parameters, working weights (if IRLS fitting method) was chosen and number of iterations taken.

Details

If method argument was set to "optim" the stats::optim function will be used to fit regression with analytically computed gradient and (minus) log likelihood functions as gr and fn arguments. Unfortunately optim does not allow for hessian to be specified. More information about how to modify optim fitting is included in controlMethod().

If method argument was set to "IRLS" the iteratively reweighted least squares. The algorithm is well know in generalised linear models. Thomas W. Yee later extended this algorithm to vector generalised linear models and in more general terms it can roughly be described as (this is Yee's description after changing some conventions):

Initialise with:

converged <- FALSE

iter <- 1

\(\boldsymbol{\beta}\) <- start

\(\boldsymbol{W}\) <- prior

\(\ell\) <- \(\ell(\boldsymbol{\beta})\)

If converged or iter > Maxiter move to step 7.

Store values from previous algorithm step:

\(\boldsymbol{W}_{-}\) <- \(\boldsymbol{W}\)

\(\ell_{-}\) <- \(\ell\)

\(\boldsymbol{\beta}_{-}\) <- \(\boldsymbol{\beta}\)

and assign values at current step:

\(\boldsymbol{\eta}\) <- \(\boldsymbol{X}_{vlm}\boldsymbol{\beta}\)

\(Z_{i}\) <- \( \eta_{i}+\frac{\partial\ell_{i}}{\partial\eta_{i}} \mathbb{E}\left(\frac{\partial^{2}\ell_{i}}{ \partial\eta_{i}^{T}\partial\eta_{i}}\right)^{-1}\)

\(\boldsymbol{W}_{ij}\) <- \(\mathbb{E}\left(\frac{\partial^{2}\ell}{ \partial\boldsymbol{\eta}_{j}^{T}\partial\boldsymbol{\eta}_{i}}\right)\)

where \(\ell_{i}\) is the ith component of log likelihood function, \(\eta_{i}\) is the vector of linear predictors associated with ith row and \(\mathbb{E}\left(\frac{\partial^{2}\ell_{i}}{ \partial\eta_{i}^{T}\partial\eta_{i}}\right)\) corresponds to weights associated with ith row and \(\boldsymbol{W}\) is a block matrix, made of diagonal matrixes \(\mathbb{E}\left(\frac{\partial^{2}\ell}{ \partial\boldsymbol{\eta}_{j}^{T}\partial\boldsymbol{\eta}_{i}}\right)\)

Regress \(\boldsymbol{Z}\) on \(\boldsymbol{X}_{vlm}\) to obtain \(\boldsymbol{\beta}\) as: \[\boldsymbol{\beta}= \left(\boldsymbol{X}_{vlm}^{T}\boldsymbol{W}\boldsymbol{X}_{vlm}\right)^{-1} \boldsymbol{X}_{vlm}^{T}\boldsymbol{W}\boldsymbol{Z}\]

Assign:

converged <- \( \ell(\boldsymbol{\beta})-\ell_{-} \( ||\boldsymbol{\beta}-\boldsymbol{\beta}_{-}||_{\infty}

iter <- iter + 1

where \(\varepsilon\) is the relative tolerance level, by default 1e-8.

Return to step 2.

Return \(\boldsymbol{\beta}, \boldsymbol{W}\), iter.

In this package we use different conventions for \(\boldsymbol{X}_{vlm}\) matrix hence slight differences are present in algorithm description but results are identical.

References

Yee, T. W. (2015). Vector Generalized Linear and Additive Models: With an Implementation in R. New York, USA: Springer. ISBN 978-1-4939-2817-0.

See also

Author

Piotr Chlebicki, Maciej Beresewicz

Examples

# \donttest{
# Get data
summary(farmsubmission)
#>    TOTAL_SUB        log_size       log_distance      C_TYPE    
#>  Min.   : 1.00   Min.   : 0.000   Min.   : 4.102   Beef :5336  
#>  1st Qu.: 1.00   1st Qu.: 4.673   1st Qu.:10.351   Dairy:6700  
#>  Median : 1.00   Median : 5.347   Median :10.778               
#>  Mean   : 2.34   Mean   : 5.259   Mean   :10.662               
#>  3rd Qu.: 3.00   3rd Qu.: 5.940   3rd Qu.:11.099               
#>  Max.   :47.00   Max.   :10.480   Max.   :12.097               

# construct vglm model matrix
X <- matrix(data = 0, nrow = 2 * NROW(farmsubmission), ncol = 7)
X[1:NROW(farmsubmission), 1:4] <- model.matrix(
  ~ 1 + log_size + log_distance + C_TYPE, 
  farmsubmission
)

X[-(1:NROW(farmsubmission)), 5:7] <- X[1:NROW(farmsubmission), c(1, 3, 4)]

# this atrrtibute tells the function which elements of the design matrix 
# correspond to which linear predictor 
attr(X, "hwm") <- c(4, 3)

# get starting points
start <- glm.fit(
  y = farmsubmission$TOTAL_SUB, 
  x = X[1:NROW(farmsubmission), 1:4], 
  family = poisson()
)$coefficients

start <- c(start, 0, 0, 0)

# call function
res <- estimatePopsize.fit(
  y = farmsubmission$TOTAL_SUB, 
  X = X, 
  method = "IRLS", 
  priorWeights = 1, 
  family = ztoigeom(), 
  control = controlMethod(verbose = 5), 
  start = start
)
#> Iteration number 1 log-likelihood: -17455.372
#> Parameter vector:  -2.255494347  0.521900283 -0.048255922  0.321168020 -1.297382847  0.049409082 -0.726587214
#> log-likelihood reduction:  1178.8778
#> Value of gradient at current step:
#>    639.53919  4035.62183  6732.72078   672.31223  -316.39394 -3358.16739  -229.50394
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 1.4296549
#> ----
#> Iteration number 2 log-likelihood: -17289.531
#> Parameter vector:  -2.859491920  0.627917105 -0.063516471  0.573159952 -2.327571214  0.074464787 -0.734988992
#> log-likelihood reduction:  165.84115
#> Value of gradient at current step:
#>    77.37365182  329.72667483  823.92206208    0.28711463  -53.27713944 -568.21808076  -29.69092406
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 1.0301884
#> ----
#> Iteration number 3 log-likelihood: -17279.272
#> Parameter vector:  -2.710874025  0.613067896 -0.069548715  0.537208696 -2.550651004  0.071488122 -0.939145580
#> log-likelihood reduction:  10.258788
#> Value of gradient at current step:
#>   35.9649654 218.1210745 385.4441808  32.8759028  -6.8429684 -71.8589804  -5.7178443
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 0.22307979
#> ----
#> Iteration number 4 log-likelihood: -17278.776
#> Parameter vector:  -2.78426085  0.61628333 -0.06440012  0.53843272 -3.10491635  0.12060422 -1.04138882
#> log-likelihood reduction:  0.49548535
#> Value of gradient at current step:
#>   1.67966432 11.13935043 16.42827781  1.15147049 -0.53787945 -5.57182317 -0.74695785
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 0.55426534
#> ----
#> Iteration number 5 log-likelihood: -17278.762
#> Parameter vector:  -2.77818073  0.61674992 -0.06504016  0.53517478 -3.10447410  0.12145802 -1.08163762
#> log-likelihood reduction:  0.014133264
#> Value of gradient at current step:
#>   0.232501334  2.204257359  2.694809270  0.131590616 -0.071608376 -0.605938875 -0.064356213
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 0.040248796
#> ----
#> Iteration number 6 log-likelihood: -17278.761
#> Parameter vector:  -2.78602286  0.61698772 -0.06441877  0.53495927 -3.19325142  0.12969021 -1.08314735
#> log-likelihood reduction:  0.00072206104
#> Value of gradient at current step:
#>  -0.0149254818 -0.0442854058 -0.2585639446 -0.0671445437 -0.0089103073 -0.1428452785 -0.0408361061
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 0.088777321
#> ----
#> Iteration number 7 log-likelihood: -17278.761
#> Parameter vector:  -2.783307402  0.617006834 -0.064659772  0.534565433 -3.160032510  0.126742033 -1.087078105
#> log-likelihood reduction:  0.000086867327
#> Value of gradient at current step:
#>   0.0219006963  0.2137543746  0.2878052106  0.0210725209 -0.0038258075 -0.0055588958  0.0088000042
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 0.033218913
#> ----
#> Iteration number 8 log-likelihood: -17278.761
#> Parameter vector:  -2.785132183  0.617029483 -0.064506904  0.534668321 -3.181900933  0.128724679 -1.085840518
#> log-likelihood reduction:  0.00002149725
#> Value of gradient at current step:
#>  -0.00917986044 -0.07823752476 -0.12764140893 -0.01528913768  0.00031447068 -0.01459257986 -0.00776040766
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 0.021868422
#> ----
#> Iteration number 9 log-likelihood: -17278.761
#> Parameter vector:  -2.784182050  0.617023943 -0.064588093  0.534586386 -3.170374939  0.127687086 -1.086727858
#> log-likelihood reduction:  0.0000064244196
#> Value of gradient at current step:
#>   0.00580810483  0.05244014329  0.07853720972  0.00778331030 -0.00057098473  0.00429221926  0.00373823938
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 0.011525994
#> ----
#> Iteration number 10 log-likelihood: -17278.761
#> Parameter vector:  -2.784725925  0.617028507 -0.064541975  0.534626993 -3.176948180  0.128280435 -1.086273279
#> log-likelihood reduction:  0.0000019945583
#> Value of gradient at current step:
#>  -0.00307033642 -0.02728421573 -0.04201045343 -0.00448214639  0.00022935424 -0.00326931977 -0.00221219810
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 0.006573241
#> ----
#> Iteration number 11 log-likelihood: -17278.761
#> Parameter vector:  -2.78442528  0.61702627 -0.06456754  0.53460327 -3.17330966  0.12795232 -1.08653540
#> log-likelihood reduction:  0.00000061988248
#> Value of gradient at current step:
#>   0.00175710834  0.01568170285  0.02390375474  0.00247668337 -0.00014882078  0.00161926853  0.00120723718
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 0.0036385179
#> ----
#> Iteration number 12 log-likelihood: -17278.761
#> Parameter vector:  -2.784593634  0.617027580 -0.064553239  0.534616288 -3.175346365  0.128136053 -1.086390890
#> log-likelihood reduction:  0.00000019314757
#> Value of gradient at current step:
#>  -0.000968762304 -0.008641334328 -0.013216142726 -0.001385713857  0.000077984442 -0.000953809209 -0.000679710368
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 0.0020367031
#> ----
#> Iteration number 13 log-likelihood: -17278.761
#> Parameter vector:  -2.784499807  0.617026860 -0.064561213  0.534608978 -3.174211159  0.128033659 -1.086471902
#> log-likelihood reduction:  0.000000060121238
#> Value of gradient at current step:
#>   0.000543879318  0.004849594593  0.007409081324  0.000772957725 -0.000044801246  0.000519451077  0.000377932793
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 0.0011352061
#> ----
#> Iteration number 14 log-likelihood: -17278.761
#> Parameter vector:  -2.784552191  0.617027265 -0.064556762  0.534613048 -3.174844941  0.128090828 -1.086426773
#> log-likelihood reduction:  0.000000018728315
#> Value of gradient at current step:
#>  -0.00030256412 -0.00269909096 -0.00412482654 -0.00043127032  0.00002466148 -0.00029328219 -0.00021122304
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 0.00063378215
#> ----
#> Iteration number 15 log-likelihood: -17278.761
#> Parameter vector:  -2.784522964  0.617027040 -0.064559245  0.534610775 -3.174491334  0.128058932 -1.086451974
#> log-likelihood reduction:  0.00000000583168
#> Value of gradient at current step:
#>   0.000169118291  0.001508137257  0.002304646966  0.000240720360 -0.000013855243  0.000162719782  0.000117791113
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 0.0003536072
#> ----
#> Value of analytically computed hessian at fitted regression coefficients:
#>             [,1]         [,2]         [,3]        [,4]       [,5]        [,6]
#> [1,]  -5533.0360  -31139.1628  -58786.3824  -3921.3952  407.46348  23473.8262
#> [2,] -31139.1628 -179723.7217 -330742.8394 -22956.5461 4362.01544   1060.0139
#> [3,] -58786.3824 -330742.8394 -627035.9660 -41504.3671  185.81715   4362.0154
#> [4,]  -3921.3952  -22956.5461  -41504.3671  -3921.3952 2193.72685  46845.5137
#> [5,]    407.4635     185.8172   46845.5137   1060.0139  -88.14760   -948.1115
#> [6,]   2193.7268    4362.0154    1971.9380   1971.9380 -948.11154 -10234.8910
#> [7,]   4362.0154   23473.8262     185.8172    185.8172  -28.97964   -312.6869
#>            [,7]
#> [1,] 1971.93804
#> [2,]  185.81715
#> [3,] 1971.93804
#> [4,]  185.81715
#> [5,]  -28.97964
#> [6,] -312.68692
#> [7,]  -28.97964
#> The matrix above has the following eigen values:
#>  -811159.7+0i -17207.51+0i -2161.919+9274.748i -2161.919-9274.748i 6057.218+0i 1581.239+0i -1513.504+0i 

# extract results

# regression coefficient vector
res$beta
#> [1] -2.78452296  0.61702704 -0.06455925  0.53461077 -3.17449133  0.12805893
#> [7] -1.08645197

# check likelihood
ll <- ztnegbin()$makeMinusLogLike(y = farmsubmission$TOTAL_SUB, X = X)

-ll(res$beta)
#> [1] -19536.01

# number of iterations
res$iter
#> [1] 15

# working weights
head(res$weights)
#>          lambda       mixed       mixed       omega
#> [1,] 0.22688312 -0.03207963 -0.03207963 0.006351956
#> [2,] 0.66642253 -0.03199578 -0.03199578 0.006238854
#> [3,] 0.08469406 -0.01202233 -0.01202233 0.001899477
#> [4,] 0.14084340 -0.01990317 -0.01990317 0.003397031
#> [5,] 0.34099080 -0.04762290 -0.04762290 0.012045206
#> [6,] 0.42290801 -0.05782236 -0.05782236 0.018478738

# Compare with optim call

res2 <- estimatePopsize.fit(
  y = farmsubmission$TOTAL_SUB, 
  X = X, 
  method = "optim", 
  priorWeights = 1, 
  family = ztoigeom(), 
  start = start, 
  control = controlMethod(verbose = 1)
)
#> iter   10 value 17321.626463
#> iter   20 value 17301.696934
#> iter   30 value 17281.887033
#> iter   40 value 17281.801823
#> iter   50 value 17281.438730
#> iter   60 value 17280.194130
#> iter   70 value 17280.188789
#> iter   80 value 17280.176569
#> iter   90 value 17280.122521
#> iter  100 value 17280.027214
#> iter  110 value 17279.476714
#> iter  120 value 17279.216352
#> iter  130 value 17279.089746
#> iter  140 value 17278.919206
#> iter  150 value 17278.894635
#> iter  160 value 17278.867207
#> iter  170 value 17278.864148
#> iter  180 value 17278.802633
#> iter  190 value 17278.776768
#> iter  200 value 17278.772490
#> iter  210 value 17278.768282
#> iter  220 value 17278.768123
#> iter  230 value 17278.767875
#> iter  240 value 17278.767809
#> iter  250 value 17278.767469
#> iter  260 value 17278.765452
#> iter  270 value 17278.761627
#> iter  280 value 17278.761313
#> iter  290 value 17278.761310
#> iter  300 value 17278.761306
#> iter  310 value 17278.761304
#> iter  320 value 17278.761298
#> iter  330 value 17278.761287
#> iter  340 value 17278.761287
#> iter  350 value 17278.761287
#> iter  360 value 17278.761287
#> iter  370 value 17278.761287
#> iter  380 value 17278.761287
#> iter  390 value 17278.761287
#> iter  400 value 17278.761287
#> iter  410 value 17278.761287
#> iter  420 value 17278.761287
#> iter  430 value 17278.761287
#> iter  440 value 17278.761287
#> iter  450 value 17278.761287
#> iter  460 value 17278.761287
#> iter  470 value 17278.761287
#> iter  480 value 17278.761287
#> iter  490 value 17278.761287
#> iter  500 value 17278.761287
#> iter  510 value 17278.761287
#> final  value 17278.761287 
#> converged
# extract results

# regression coefficient vector
res2$beta
#> [1] -2.78453076  0.61702707 -0.06455857  0.53461150 -3.17461277  0.12806986
#> [7] -1.08644213

# check likelihood
-ll(res2$beta)
#> [1] -19536.01

# number of calls to log lik function
# since optim does not return the number of
# iterations
res2$iter
#> function gradient 
#>      616      616 

# optim does not calculated working weights
head(res2$weights)
#> [1] 1
# }