Regression fitting in single source capture-recapture models

estimatePopsize.fit does for estimatePopsize what glm.fit does for glm. It is internally called in estimatePopsize. Since estimatePopsize does much more than just regression fitting estimatePopsize.fit is much faster.

Usage

estimatePopsize.fit(y, X, family, control, method, priorWeights, start, ...)

Arguments

y

vector of dependent variables.

X

model matrix, the vglm one.

family

same as model in estimatePopsize.

control

control parameters created in controlModel.

method

method of estimation same as in estimatePopsize.

priorWeights

vector of prior weights its the same argument as weights in estimatePopsize.

start

initial value of regression parameters.

...

arguments to pass to other methods.

Value

List with regression parameters, working weights (if IRLS fitting method) was chosen and number of iterations taken.

Details

If method argument was set to "optim" the stats::optim function will be used to fit regression with analyticly computed gradient and (minus) log likelihood functions as gr and fn arguments. Unfortunately optim does not allow for hessian to be specified. More information about how to modify optim fitting is included in controlMethod().

If method argument was set to "IRLS" the iteratively reweighted least squares. The algorithm is well know in generalised linear models. Thomas W. Yee later extended this algorithm to vector generalised linear models and in more general terms it can roughly be described as (this is Yee's description after changing some conventions):

Initialise with:

converged <- FALSE

iter <- 1

\(\boldsymbol{\beta}\) <- start

\(\boldsymbol{W}\) <- prior

\(\ell\) <- \(\ell(\boldsymbol{\beta})\)

If converged or iter > Maxiter move to step 7.

Store values from previous algorithm step:

\(\boldsymbol{W}_{-}\) <- \(\boldsymbol{W}\)

\(\ell_{-}\) <- \(\ell\)

\(\boldsymbol{\beta}_{-}\) <- \(\boldsymbol{\beta}\)

and assign values at current step:

\(\boldsymbol{\eta}\) <- \(\boldsymbol{X}_{vlm}\boldsymbol{\beta}\)

\(Z_{i}\) <- \(\eta_{i}+\frac{\partial\ell_{i}}{\partial\eta_{i}} \mathbb{E}\left(\frac{\partial^{2}\ell_{i}}{ \partial\eta_{i}^{T}\partial\eta_{i}}\right)^{-1}\)

\(\boldsymbol{W}_{ij}\) <- \(\mathbb{E}\left(\frac{\partial^{2}\ell}{ \partial\boldsymbol{\eta}_{j}^{T}\partial\boldsymbol{\eta}_{i}}\right)\)

where \(\ell_{i}\) is the ith component of log likelihood function, \(\eta_{i}\) is the vector of linear predictors associated with ith row and \(\mathbb{E}\left(\frac{\partial^{2}\ell_{i}}{\partial\eta_{i}^{T}\partial\eta_{i}}\right)\) corresponds to weights associated with ith row and \(\boldsymbol{W}\) is a block matrix, made of diagonal matrixes \(\mathbb{E}\left(\frac{\partial^{2}\ell}{ \partial\boldsymbol{\eta}_{j}^{T}\partial\boldsymbol{\eta}_{i}}\right)\)

Regress \(\boldsymbol{Z}\) on \(\boldsymbol{X}_{vlm}\) to obtain \(\boldsymbol{\beta}\) as: \[\boldsymbol{\beta}= \left(\boldsymbol{X}_{vlm}^{T}\boldsymbol{W}\boldsymbol{X}_{vlm}\right)^{-1} \boldsymbol{X}_{vlm}^{T}\boldsymbol{W}\boldsymbol{Z}\]

Assign:

converged <- \(\ell(\boldsymbol{\beta})-\ell_{-} \lt \varepsilon \ell_{-}\) \(\lor\) \(||\boldsymbol{\beta}-\boldsymbol{\beta}_{-}||_{\infty} \lt \varepsilon\)

iter <- iter + 1

where \(\varepsilon\) is the relative tolerance level, by default 1e-8.

Return to step 2.

Return \(\boldsymbol{\beta}, \boldsymbol{W}\), iter.

In this package we use different conventions for \(\boldsymbol{X}_{vlm}\) matrix hence slight differences are present in algorithm description but results are identical.

References

Yee, T. W. (2015). Vector Generalized Linear and Additive Models: With an Implementation in R. New York, USA: Springer. ISBN 978-1-4939-2817-0.

See also

Author

Piotr Chlebicki, Maciej Beręsewicz

Examples

# Get data
summary(farmsubmission)
#>    TOTAL_SUB        log_size       log_distance      C_TYPE    
#>  Min.   : 1.00   Min.   : 0.000   Min.   : 4.102   Beef :5336  
#>  1st Qu.: 1.00   1st Qu.: 4.673   1st Qu.:10.351   Dairy:6700  
#>  Median : 1.00   Median : 5.347   Median :10.778               
#>  Mean   : 2.34   Mean   : 5.259   Mean   :10.662               
#>  3rd Qu.: 3.00   3rd Qu.: 5.940   3rd Qu.:11.099               
#>  Max.   :47.00   Max.   :10.480   Max.   :12.097               

# construct vglm model matrix
X <- matrix(data = 0, nrow = 2 * NROW(farmsubmission), ncol = 7)
X[1:NROW(farmsubmission), 1:4] <- model.matrix(~ 1 + log_size + log_distance + C_TYPE, 
                                               farmsubmission)
X[-(1:NROW(farmsubmission)), 5:7] <- X[1:NROW(farmsubmission), c(1, 3, 4)]
# this atrrtibute tells the function which elements of the design matrix 
# correspond to which linear predictor 
attr(X, "hwm") <- c(4, 3)

# get starting points
start <- glm.fit(y = farmsubmission$TOTAL_SUB, 
                 x = X[1:NROW(farmsubmission), 1:4], 
                 family = poisson())$coefficients
start <- c(start, start)[-6]

# call function
res <- estimatePopsize.fit(
  y = farmsubmission$TOTAL_SUB, X = X, 
  method = "IRLS", 
  priorWeights = 1, 
  family = ztnegbin(), 
  control = controlMethod(verbose = 5, stepsize = .75), 
  start = start
)
#> Iteration number 1 log-likelihood: -17823.492
#> Parameter vector:  -1.863614075  0.475204651 -0.059013713  0.508672480 -0.327870976  0.086003423 -0.412206923
#> log-likelihood reduction:  901.05075
#> Value of gradient at current step:
#>   -1990.34014 -10046.54007 -21321.73120   -881.45386   -716.47130  -7710.75888   -223.33219
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 1.0377746
#> ----
#> Iteration number 2 log-likelihood: -17437.868
#> Parameter vector:  -2.83898496  0.57720331 -0.10523289  1.12721800  0.56681332  0.10483162 -1.23066712
#> log-likelihood reduction:  385.62401
#> Value of gradient at current step:
#>    -952.47828  -4741.57396 -10212.18518   -343.98244   -618.99520  -6663.80140   -154.63239
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 0.97537088
#> ----
#> Iteration number 3 log-likelihood: -17295.872
#> Parameter vector:  -4.277460156  0.634940802 -0.094367509  2.034789937  2.128998318  0.045964537 -2.151291285
#> log-likelihood reduction:  141.99625
#> Value of gradient at current step:
#>   -413.851502 -2040.578657 -4436.735470  -109.135451  -335.376402 -3612.781025   -58.889278
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 1.562185
#> ----
#> Iteration number 4 log-likelihood: -17254.782
#> Parameter vector:  -5.072408286  0.657156480 -0.065245482  2.381886041  2.961574999 -0.012066361 -2.398344381
#> log-likelihood reduction:  41.090197
#> Value of gradient at current step:
#>   -139.197813  -687.872162 -1494.327940   -31.450516  -119.173553 -1288.504052   -18.285250
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 0.83257668
#> ----
#> Iteration number 5 log-likelihood: -17249.508
#> Parameter vector:  -4.785542658  0.662080248 -0.044539312  1.847021276  2.738430184 -0.051852315 -1.769105789
#> log-likelihood reduction:  5.2739327
#> Value of gradient at current step:
#>   -62.5681200 -316.4732532 -673.2851829   -8.2755289  -52.4549093 -567.4986454   -4.9417204
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 0.62923859
#> ----
#> Iteration number 6 log-likelihood: -17248.248
#> Parameter vector:  -4.838358102  0.662181992 -0.035679178  1.805785896  2.842862541 -0.069778744 -1.688871750
#> log-likelihood reduction:  1.2603975
#> Value of gradient at current step:
#>   -18.0169421  -92.8169585 -193.8385273   -1.8798049  -15.1781828 -164.9037392   -1.1047930
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 0.10443236
#> ----
#> Iteration number 7 log-likelihood: -17248.124
#> Parameter vector:  -4.856424279  0.661637750 -0.030877664  1.776299716  2.894932574 -0.079055152 -1.644112756
#> log-likelihood reduction:  0.12375469
#> Value of gradient at current step:
#>   -5.06601911 -26.72531290 -54.37487852  -0.31528393  -4.27867262 -46.73618805  -0.15560324
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 0.052070032
#> ----
#> Iteration number 8 log-likelihood: -17248.111
#> Parameter vector:  -4.869878276  0.661294400 -0.028509607  1.766665256  2.926220022 -0.083411593 -1.629501617
#> log-likelihood reduction:  0.012599044
#> Value of gradient at current step:
#>   -1.3791306185  -7.4663980661 -14.7133524835  -0.0027426914  -1.1676544673 -12.8743485954   0.0215439672
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 0.031287448
#> ----
#> Iteration number 9 log-likelihood: -17248.11
#> Parameter vector:  -4.878959253  0.661144077 -0.027331378  1.764109284  2.944258141 -0.085493615 -1.625463512
#> log-likelihood reduction:  0.0014672115
#> Value of gradient at current step:
#>  -0.368913065 -2.048699268 -3.884684644  0.031430474 -0.312178145 -3.504357533  0.030488340
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 0.018038119
#> ----
#> Iteration number 10 log-likelihood: -17248.109
#> Parameter vector:  -4.884625375  0.661089164 -0.026731658  1.763718234  2.954460230 -0.086525598 -1.624686621
#> log-likelihood reduction:  0.00022648247
#> Value of gradient at current step:
#>  -0.096875744 -0.548122695 -0.991559296  0.020795430 -0.081408996 -0.947603064  0.017228273
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 0.010202089
#> ----
#> Iteration number 11 log-likelihood: -17248.109
#> Parameter vector:  -4.888000502  0.661072319 -0.026419129  1.763867036  2.960181751 -0.087055092 -1.624768843
#> log-likelihood reduction:  0.000047774331
#> Value of gradient at current step:
#>  -0.0244397651 -0.1376704317 -0.2338731302  0.0104103315 -0.0201292559 -0.2533524486  0.0078301405
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 0.0057215212
#> ----
#> Iteration number 12 log-likelihood: -17248.109
#> Parameter vector:  -4.889956420  0.661068670 -0.026253010  1.764077434  2.963376789 -0.087334144 -1.624992315
#> log-likelihood reduction:  0.000012449145
#> Value of gradient at current step:
#>  -0.0055313064 -0.0285989665 -0.0431543682  0.0047561975 -0.0043078410 -0.0657477152  0.0032151064
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 0.0031950377
#> ----
#> Iteration number 13 log-likelihood: -17248.109
#> Parameter vector:  -4.891071396  0.661068881 -0.026163395  1.764238408  2.965156867 -0.087484001 -1.625177606
#> log-likelihood reduction:  0.0000035915436
#> Value of gradient at current step:
#>  -0.00084286134 -0.00187662677  0.00006712155  0.00211608014 -0.00049470766 -0.01574614830  0.00124263181
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 0.0017800781
#> ----
#> Iteration number 14 log-likelihood: -17248.109
#> Parameter vector:  -4.891700725  0.661069844 -0.026114548  1.764342988  2.966147382 -0.087565490 -1.625301537
#> log-likelihood reduction:  0.0000010801996
#> Value of gradient at current step:
#>   0.00015910985  0.00314638510  0.00679966516  0.00094773468  0.00025496115 -0.00292540930  0.00045995750
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 0.00099051501
#> ----
#> Iteration number 15 log-likelihood: -17248.109
#> Parameter vector:  -4.892053814  0.661070692 -0.026087737  1.764406292  2.966698186 -0.087610160 -1.625377602
#> log-likelihood reduction:  0.00000033018296
#> Value of gradient at current step:
#>  0.000267949697 0.003028308015 0.005713247289 0.000435422817 0.000291345164 0.000027911184 0.000163958646
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 0.00055080374
#> ----
#> Iteration number 16 log-likelihood: -17248.109
#> Parameter vector:  -4.892251193  0.661071277 -0.026072955  1.764443253  2.967004372 -0.087634771 -1.625422349
#> log-likelihood reduction:  0.00000010155782
#> Value of gradient at current step:
#>  0.000201240710 0.002082281200 0.003738961046 0.000207088974 0.000204770704 0.000491982940 0.000056070532
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 0.0003061859
#> ----
#> Iteration number 17 log-likelihood: -17248.109
#> Parameter vector:  -4.892361281  0.661071643 -0.026064781  1.764464407  2.967174548 -0.087648375 -1.625448069
#> log-likelihood reduction:  0.000000031312084
#> Value of gradient at current step:
#>  0.000127397155 0.001284304865 0.002245575651 0.000102112246 0.000126264478 0.000412962287 0.000018089806
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 0.00017017649
#> ----
#> Iteration number 18 log-likelihood: -17248.109
#> Parameter vector:  -4.892422599  0.661071861 -0.026060253  1.764476375  2.967269123 -0.087655909 -1.625462658
#> log-likelihood reduction:  0.0000000096661097
#> Value of gradient at current step:
#>  0.0000755209312 0.0007549148594 0.0012988453582 0.0000520085308 0.0000738565809 0.0002710898226 0.0000052602827
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 0.000094575234
#> ----
#> Value of analytically computed hessian at fitted regression coefficients:
#>            [,1]        [,2]        [,3]       [,4]        [,5]        [,6]
#> [1,]  -4571.199  -25603.055  -48636.535  -3380.822  -2509.0413 -145504.055
#> [2,] -25603.055 -147121.914 -272342.907 -19693.777 -26719.7905   -8691.146
#> [3,] -48636.535 -272342.907 -519445.459 -35856.227  -1501.4754  -26719.790
#> [4,]  -3380.822  -19693.777  -35856.227  -3380.822 -13666.9336 -285668.710
#> [5,]  -2509.041   -1501.475 -285668.710  -8691.146  -1717.8462  -18309.468
#> [6,] -13666.934  -26719.790  -15908.320 -15908.320 -18309.4675 -195877.421
#> [7,] -26719.790 -145504.055   -1501.475  -1501.475   -860.5633   -9111.942
#>             [,7]
#> [1,] -15908.3204
#> [2,]  -1501.4754
#> [3,] -15908.3204
#> [4,]  -1501.4754
#> [5,]   -860.5633
#> [6,]  -9111.9423
#> [7,]   -860.5633
#> The matrix above has the following eigen values:
#>  -681851.2+0i -215719.6+0i -2322.41+58791.38i -2322.41-58791.38i 25652.64+0i 10560.49+0i -6972.756+0i 

# extract results

# regression coefficient vector
res$beta
#> [1] -4.89242260  0.66107186 -0.02606025  1.76447637  2.96726912 -0.08765591
#> [7] -1.62546266

# check likelihood
ll <- ztnegbin()$makeMinusLogLike(y = farmsubmission$TOTAL_SUB, X = X)

-ll(res$beta)
#> [1] -17248.11

# number of iterations
res$iter
#> [1] 18

# working weights
head(res$weights)
#>          lambda      mixed      mixed      alpha
#> [1,] 0.19146547 0.16520177 0.16520177 0.14282624
#> [2,] 0.57006082 0.23023980 0.23023980 0.13334928
#> [3,] 0.07855612 0.06890872 0.06890872 0.06047553
#> [4,] 0.12737155 0.11112052 0.11112052 0.09703535
#> [5,] 0.25529253 0.21624875 0.21624875 0.18398582
#> [6,] 0.28665185 0.23845695 0.23845695 0.19987565

# Compare with optim call

res2 <- estimatePopsize.fit(
  y = farmsubmission$TOTAL_SUB, X = X, 
  method = "optim", 
  priorWeights = 1, 
  family = ztnegbin(), 
  start = start, 
  control = controlMethod()
)
# extract results

# regression coefficient vector
res2$beta
#> [1] -4.89254268  0.66107233 -0.02604925  1.76447646  2.96745272 -0.08767306
#> [7] -1.62546335

# check likelihood
-ll(res2$beta)
#> [1] -17248.11

# number of calls to log lik function
# since optim does not return the number of
# iterations
res2$iter
#> function gradient 
#>     1003     1003 

# optim does not calculated working weights
head(res2$weights)
#> [1] 1