Regression fitting in single source capture-recapture models

estimatePopsize.fit does for estimatePopsize what glm.fit does for glm. It is internally called in estimatePopsize. Since estimatePopsize does much more than just regression fitting estimatePopsize.fit is much faster.

Usage

estimatePopsize.fit(y, X, family, control, method, priorWeights, start, ...)

Arguments

y

vector of dependent variables.

X

model matrix, the vglm one.

family

same as model in estimatePopsize.

control

control parameters created in controlModel.

method

method of estimation same as in estimatePopsize.

priorWeights

vector of prior weights its the same argument as weights in estimatePopsize.

start

initial value of regression parameters.

...

arguments to pass to other methods.

Value

List with regression parameters, working weights (if IRLS fitting method) was chosen and number of iterations taken.

Details

If method argument was set to "optim" the stats::optim function will be used to fit regression with analyticly computed gradient and (minus) log likelihood functions as gr and fn arguments. Unfortunately optim does not allow for hessian to be specified. More information about how to modify optim fitting is included in controlMethod().

If method argument was set to "IRLS" the iteratively reweighted least squares. The algorithm is well know in generalised linear models. Thomas W. Yee later extended this algorithm to vector generalised linear models and in more general terms it can roughly be described as (this is Yee's description after changing some conventions):

Initialise with:

converged <- FALSE

iter <- 1

\(\boldsymbol{\beta}\) <- start

\(\boldsymbol{W}\) <- prior

\(\ell\) <- \(\ell(\boldsymbol{\beta})\)

If converged or iter > Maxiter move to step 7.

Store values from previous algorithm step:

\(\boldsymbol{W}_{-}\) <- \(\boldsymbol{W}\)

\(\ell_{-}\) <- \(\ell\)

\(\boldsymbol{\beta}_{-}\) <- \(\boldsymbol{\beta}\)

and assign values at current step:

\(\boldsymbol{\eta}\) <- \(\boldsymbol{X}_{vlm}\boldsymbol{\beta}\)

\(Z_{i}\) <- \(\eta_{i}+\frac{\partial\ell_{i}}{\partial\eta_{i}}\mathbb{E}\left(\frac{\partial^{2}\ell_{i}}{\partial\eta_{i}^{T}\partial\eta_{i}}\right)^{-1}\)

\(\boldsymbol{W}_{ij}\) <- \(\mathbb{E}\left(\frac{\partial^{2}\ell}{\partial\boldsymbol{\eta}_{j}^{T}\partial\boldsymbol{\eta}_{i}}\right)\)

where \(\ell_{i}\) is the ith component of log likelihood function, \(\eta_{i}\) is the vector of linear predictors associated with ith row and \(\mathbb{E}\left(\frac{\partial^{2}\ell_{i}}{\partial\eta_{i}^{T}\partial\eta_{i}}\right)\) corresponds to weights associated with ith row and \(\boldsymbol{W}\) is a block matrix, made of diagonal matrixes \(\mathbb{E}\left(\frac{\partial^{2}\ell}{\partial\boldsymbol{\eta}_{j}^{T}\partial\boldsymbol{\eta}_{i}}\right)\).

Regress \(\boldsymbol{Z}\) on \(\boldsymbol{X}_{vlm}\) to obtain \(\boldsymbol{\beta}\) as: \[\boldsymbol{\beta}=\left(\boldsymbol{X}_{vlm}^{T}\boldsymbol{W}\boldsymbol{X}_{vlm}\right)^{-1}\boldsymbol{X}_{vlm}^{T}\boldsymbol{W}\boldsymbol{Z}\]

Assign:

converged <- \(\ell(\boldsymbol{\beta})-\ell_{-} or \(||\boldsymbol{\beta}-\boldsymbol{\beta}_{-}||_{\infty}

iter <- iter + 1

where \(\varepsilon\) is the relative tolerance level, by default 1e-8.

Return to step 2.

Return \(\boldsymbol{\beta}, \boldsymbol{W}\), iter.

In this package we use different conventions for \(\boldsymbol{X}_vlm\) matrix hence slight differences are present in algorithm description but results are identical.

References

Yee, T. W. (2015). Vector Generalized Linear and Additive Models: With an Implementation in R. New York, USA: Springer. ISBN 978-1-4939-2817-0.

See also

Author

Piotr Chlebicki, Maciej Beręsewicz

Examples

# Set seed for reproducibility
set.seed(123)

# Set population size and 'true' regression coefficients
N <- 7562
beta1 <- c(1, .2, -.95, -3)
beta2 <- c(2, 0, -.01, .3)

# Generate covariates
x1 <- rpois(n = N, lambda = 1.85)
x2 <- rnorm(n = N)
x3 <- rbinom(n = N, size = 1, prob = runif(n = 1))

# Compute linear predictors
eta1 <- cbind(1, x1, x2, x3) %*% beta1
eta2 <- cbind(1, x1, x2, x3) %*% beta2

# generate dependent variable
y <- ztoipoisson()$simulate(n = N, eta = cbind(eta1, eta2))
# optionally visualise destribution on y with barplot
# barplot(table(y))
# truncate data
x1 <- x1[y>0]
x2 <- x2[y>0]
x3 <- x3[y>0]
y <- y[y>0]

# construct vglm model matrix
X <- matrix(data = 0, nrow = 2 * length(y), ncol = 8)
X[1:length(y), 1:4] <- model.matrix(~ 1 + x1 + x2 + x3)
X[-(1:length(y)), 5:8] <- X[1:length(y), 1:4]
# this arrtibute tells the function which elements of the design matrix 
# correspond to which linear predictor 
attr(X, "hwm") <- c(4, 4)

# finally call the function
beta <- c(beta1, beta2)

res <- estimatePopsize.fit(
  y = y, X = X, 
  method = "IRLS", 
  priorWeights = 1, 
  family = ztoipoisson(), 
  control = controlMethod(verbose = 5), 
  start = beta + rnorm(n = length(beta), sd = .1)
)
#> Iteration number 1 log-likelihood: -24744.076
#> Parameter vector:  -10.32608769   0.58037475  -1.33218497   8.36264342  -4.15462875   1.00033619  -1.63224339   5.90341582
#> log-likelihood reduction:  -22645.89
#> Value of gradient at current step:
#>   2008.841853  3991.135030 -2052.038101  -512.014214  -163.400825  -465.615846   178.933201   -78.624749
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 11.372388
#> Taking a modified step....
#> Iteration number 1 log-likelihood: -2007.6174
#> Parameter vector:   0.69091287  0.26326159 -1.03420854 -2.59476489  1.84076445  0.12091652 -0.12884159  0.31561786
#> log-likelihood reduction:  90.568444
#> Value of gradient at current step:
#>   174.865107  -26.513615  158.337969  -10.824417 -233.596320 -553.952664  129.236794  -26.977372
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 0.35538711
#> ----
#> Iteration number 2 log-likelihood: -24826.724
#> Parameter vector:  -10.32648783   0.58069155  -1.33278511   8.44612934  -4.15796616   1.00010731  -1.62897813   6.40259530
#> log-likelihood reduction:  -22819.107
#> Value of gradient at current step:
#>   1952.014020  3736.765638 -1936.872567  -568.841689  -167.743878  -471.534666   180.607498   -83.197848
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 11.040894
#> Taking a modified step....
#> Iteration number 2 log-likelihood: -2007.6174
#> Parameter vector:   0.69091287  0.26326159 -1.03420854 -2.59476489  1.84076445  0.12091652 -0.12884159  0.31561786
#> log-likelihood reduction:  0.00000000000022737368
#> Value of gradient at current step:
#>   174.865107  -26.513615  158.337969  -10.824417 -233.596320 -553.952664  129.236794  -26.977372
#> Algorithm will terminate if one of following conditions will be met:
#> The increase to minus log-likelihood will be bellow chosen value of epsilon 1e-08 
#> Maximum change to the vector of regression parameters will be bellow the chosen value of epsilon.
#> At current step the highest change was: 0.00000000000000044408921
#> ----
#> Warning: Convergence at halfstepsize
#> Value of analytically computed hessian at fitted regression coefficients:
#>             [,1]         [,2]        [,3]       [,4]       [,5]       [,6]
#> [1,] -2657.85992  -7217.84982  3584.50587 -142.78323   77.04236  143.27686
#> [2,] -7217.84982 -27619.89450 10288.10240 -417.13512  143.27686  407.88203
#> [3,]  3584.50587  10288.10240 -7892.62858  223.19784  -27.47559  -46.63509
#> [4,]  -142.78323   -417.13512   223.19784 -142.78323   48.89458  101.21679
#> [5,]    77.04236    143.27686   -27.47559   48.89458 -151.58136 -298.78820
#> [6,]   143.27686    407.88203   -46.63509  101.21679 -298.78820 -876.15563
#> [7,]   -27.47559    -46.63509    83.26395  -42.37576   80.04882  147.80742
#> [8,]    48.89458    101.21679   -42.37576   48.89458  -58.24756 -128.05861
#>            [,7]       [,8]
#> [1,]  -27.47559   48.89458
#> [2,]  -46.63509  101.21679
#> [3,]   83.26395  -42.37576
#> [4,]  -42.37576   48.89458
#> [5,]   80.04882  -58.24756
#> [6,]  147.80742 -128.05861
#> [7,] -177.00319   60.28389
#> [8,]   60.28389  -58.24756
#> The matrix above has the following eigen values:
#>  -20.52098 -36.83832 -110.5987 -171.7599 -520.8088 -1054.241 -3577.607 -34083.78 

# extract results

# regression coefficient vector
res$beta
#> [1]  0.6909129  0.2632616 -1.0342085 -2.5947649  1.8407645  0.1209165 -0.1288416
#> [8]  0.3156179

# compare with actual true coeffictient vector
beta
#> [1]  1.00  0.20 -0.95 -3.00  2.00  0.00 -0.01  0.30

# number of iterations
res$iter
#> [1] 2

# working weights
head(res$weights)
#>            lambda         mixed         mixed        omega
#> [1,] 5.701175e-03 -4.935010e-03 -4.935010e-03 4.890802e-03
#> [2,] 5.418062e-04 -1.761896e-06 -1.761896e-06 1.119346e-03
#> [3,] 1.775422e-05 -5.795365e+02 -5.795365e+02 2.358105e-09
#> [4,] 1.918954e-04 -9.317753e-19 -9.317753e-19 1.918738e-04
#> [5,] 7.500023e-04 -3.255879e-05 -3.255879e-05 1.244941e-03
#> [6,] 1.107066e-03 -2.928194e-05 -2.928194e-05 2.030523e-03

# call with another method

res <- estimatePopsize.fit(
  y = y, X = X, 
  method = "optim", 
  priorWeights = 1, 
  family = ztoipoisson(), 
  start = beta + rnorm(n = length(beta), sd = .1), 
  control = controlMethod()
)
# extract results

# regression coefficient vector
res$beta
#> [1]  1.0047373  0.2091686 -0.9434504 -2.9622529  1.1099858 -0.1179091  0.2076010
#> [8]  1.3492574

# compare with actual true coeffictient vector
beta
#> [1]  1.00  0.20 -0.95 -3.00  2.00  0.00 -0.01  0.30

# number of iterations
res$iter
#> function gradient 
#>      132      132 

# optim does not calculated working weights
head(res$weights)
#> [1] 1